+++ mathics/builtin/drawing/uniform_polyhedra.py
@@ -14,17 +14,19 @@ from mathics.core.expression import Expression
 uniform_polyhedra_names = "tetrahedron, octahedron, dodecahedron, icosahedron"
 uniform_polyhedra_set = frozenset(uniform_polyhedra_names.split(", "))
 
+
 class UniformPolyhedron(Builtin):
     """
     <dl>
       <dt>'UniformPolyhedron["name"]'
-      <dd>is a uniform polyhedron with the given name.
+      <dd>return a uniform polyhedron with the given name.
+      <dd>Names are "tetrahedron", "octahedron", "dodecahedron", or "icosahedron".
     </dl>
 
-    >> Graphics3D[UniformPolyhedron["tetrahedron"]
+    >> Graphics3D[UniformPolyhedron["tetrahedron"]]
      = -Graphics3D-
 
-    >> Graphics3D[UniformPolyhedron["octahedron"]
+    >> Graphics3D[UniformPolyhedron["octahedron"]]
      = -Graphics3D-
     """
 
@@ -32,11 +34,38 @@ class UniformPolyhedron(Builtin):
         "argtype": f"Argument `1` is not one of: {uniform_polyhedra_names}",
     }
 
-
     def apply_with_name(self, name, evaluation):
         "UniformPolyhedron[name_String]"
 
         return Expression("UniformPolyhedron", name)
+
+
+class Dodecahedron(Builtin):
+    """
+    <dl>
+      <dt>'Dodecahedron[]'
+      <dd>a regular dodecahedron centered at the origin with unit edge length.
+    </dl>
+
+    >> Graphics3D[Dodecahedron[]]
+     = -Graphics3D-
+    """
+
+    rules = {"Dodecahedron[]": """UniformPolyhedron["dodecahedron"]"""}
+
+
+class Octahedron(Builtin):
+    """
+    <dl>
+      <dt>'Octahedron[]'
+      <dd>a regular octahedron centered at the origin with unit edge length.
+    </dl>
+
+    >> Graphics3D[Octahedron[]]
+     = -Graphics3D-
+    """
+
+    rules = {"Octahedron[]": """UniformPolyhedron["octahedron"]"""}
 
 
 class Tetrahedron(Builtin):
@@ -49,4 +78,5 @@ class Tetrahedron(Builtin):
     >> Graphics3D[Tetrahedraon[]]
      = -Graphics3D-
     """
+
     rules = {"Tetrahedron[]": """UniformPolyhedron["tetrahedron"]"""}