+++ mathics/format/json.py
@@ -216,13 +216,17 @@ def uniform_polyhedron_3d_box(self) -> list:
     face_color = self.face_color
     if face_color is not None:
         face_color = face_color.to_js()
-    data = convert_coord_collection(
-        [0, 0, 0],
-        "sphere",  #
-        face_color,
-        {"faceColor": face_color, "radius": 1},
-    )
-    print("### json UniformPolyhedron3DBox", data)
+    data = [
+        {
+            "type": "uniformPolyhedron",
+            "color": face_color[:3],
+            "faceColor": face_color,
+            "coords": [[[0, 0, 0]]],
+            "subType": self.sub_type,
+        }
+    ]
+    # print("### json UniformPolyhedron3DBox", data)
     return data
+
 
 add_conversion_fn(UniformPolyhedron3DBox, uniform_polyhedron_3d_box)